move cc_shared_state
