+++ protocol/whist/network/congestion_control/cc_shared_state.cpp
@@ -1,0 +1,3 @@
+#include "cc_shared_state.h"
+
+CCSharedState cc_shared_state;
