+++ protocol/whist/network/congestion_control/common_fixes.cpp
@@ -3,9 +3,6 @@
 #include "common_fixes.h"
 #include <sstream>
 
-
-CCSharedState cc_shared_state;
-
 namespace webrtc {
 
 std::string ToString(TimeDelta value) {
