+++ protocol/whist/network/congestion_control/common_fixes.h
@@ -17,22 +17,6 @@ extern "C"
 #include "whist/utils/clock.h"
 }
 
-struct CCSharedState
-{
-    bool g_in_slow_increase= false;
-    const double kWhistClampMin= 5.f;
-    const double g_startup_duration=6;
-    const double g_increase_ratio=0.12;
-    double max_bitrate=-1;
-    double current_bitrate_ratio=1;
-
-    double ack_bitrate= -1;
-
-};
-
-extern CCSharedState cc_shared_state;
-
-
 #define ENABLE_WHIST_CHANGE true
 namespace webrtc
 {
