+++ protocol/whist/network/congestion_control/modules/congestion_controller/goog_cc/send_side_bandwidth_estimation.cc
@@ -17,6 +17,7 @@
 #include <string>
 
 #include "common_fixes.h"
+#include "cc_shared_state.h"
 //#include "absl/strings/match.h"
 #include "api/field_trials_view.h"
 #include "api/network_state_predictor.h"
